+++ src/main/java/com/singh/astha/notification/service/security/JwtFilter.java
@@ -1,7 +1,7 @@
-package com.singh.astha.notification.service.config;
+package com.singh.astha.notification.service.security;
 
-import com.singh.astha.notification.service.dtos.internal.JwtPayload;
-import com.singh.astha.notification.service.service.JwtService;
+import com.singh.astha.notification.service.security.dtos.JwtPayload;
+import com.singh.astha.notification.service.security.service.JwtService;
 import com.singh.astha.notification.service.utils.Constants;
 import com.singh.astha.notification.service.utils.MessageConstants;
 import jakarta.servlet.FilterChain;
