+++ src/main/java/com/singh/astha/notification/service/security/service/JwtService.java
@@ -1,0 +1,8 @@
+package com.singh.astha.notification.service.security.service;
+
+import com.singh.astha.notification.service.security.dtos.JwtPayload;
+
+public interface JwtService {
+
+    JwtPayload verifyAndDecodeToken(String authHeader);
+}
