+++ src/main/java/com/singh/astha/notification/service/security/service/impl/JwtServiceImpl.java
@@ -1,12 +1,12 @@
-package com.singh.astha.notification.service.service.impl;
+package com.singh.astha.notification.service.security.service.impl;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.singh.astha.notification.service.dtos.internal.JwtPayload;
 import com.singh.astha.notification.service.exceptions.ResponseException;
-import com.singh.astha.notification.service.service.JwtService;
+import com.singh.astha.notification.service.security.dtos.JwtPayload;
+import com.singh.astha.notification.service.security.service.JwtService;
 import com.singh.astha.notification.service.utils.AppProperties;
 import com.singh.astha.notification.service.utils.Constants;
 import com.singh.astha.notification.service.utils.ErrorMessages;
