--- src/main/java/com/singh/astha/notification/service/service/JwtService.java
@@ -1,8 +1,0 @@
-package com.singh.astha.notification.service.service;
-
-import com.singh.astha.notification.service.dtos.internal.JwtPayload;
-
-public interface JwtService {
-
-    JwtPayload verifyAndDecodeToken(String authHeader);
-}
